Rename from 'inputs' to 'form'.
