+++ src/page-modules/contact/components/index.tsx
@@ -10,5 +10,5 @@ export {
   SearchableSelect,
   getLineOptions,
   getStopOptions,
-} from './inputs';
+} from './form';
 export { SectionCard } from './section-card';
